Restyle v2
Summary:
This revamps all of the css for changes based on min's designs. There was a LOT of code to change, so think of this more as a first draft. Once its landed, we'll look at it with fresh eyes and figure out what further changes to make. We also need to do polish: adjust spacing, etc.

{F230206}
{F230207}
{F230208}
{F230209}
{F230210}

Test Plan: hit local pages

Reviewers: minming, haoyi

Reviewed By: haoyi

Subscribers: haoyi, changesbot, vishal, angie, mkedia

Differential Revision: https://tails.corp.dropbox.com/D132547