Add support for replacing JobSteps that have infra failures.
Summary:
Currently only enabled for Mesos builds.

Adds replacement_id column which is used for JobSteps that are retried
as a new JobStep. If a JobStep has non-null replacement_id, it is ignored
for the purposes of a Job/JobPhase's aggregated result.

Also adds methods to sync_job and to buildstep classes to decide if a jobstep should be retried, and then to retry them.

Test Plan: Added tests and tested on local vm. Next step is testing on changes-staging.

Reviewers: kylec

Reviewed By: kylec

Subscribers: paulruan, haoyi, changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D134823