Make XUnit xml parse failing a warning; include more info
Summary:
Failure to parse XML might be our fault, but probably is the result of a bad build
script or truncation along the line. It doesn't signal a failure in Changes, therefore
it should be a warning.
Also added some other useful info to help track down whose issue it might be.

Test Plan: Unit doesn't quite seem worth it, but willing to be convinced.

Reviewers: paulruan

Reviewed By: paulruan

Subscribers: changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D138537