Store the snapshot image in jobplan.
Summary:
This associates a snapshot with a job.
It'll allow us to display what snapshot was used for a job.
Also, instead of computing the snapshot image that should be used
in the jobstep details route, it is done during jobplan creation time.
This should also keep the snapshot image used consistent
for the jobsteps of a single job.

Test Plan:
Pass tests.
Updated tests for jobstep details.
Moved tests for getting the current image.

Reviewers: jukka, kylec

Reviewed By: kylec

Subscribers: changesbot, mkedia, nate

Differential Revision: https://tails.corp.dropbox.com/D141057