Fix rerun tally if test both fails and errors
Summary: If py.test reports both an `<error>` and a `<failure>` for a given test, the fact that both of them say `rerun="3"` does not mean that the test ran 6 times. It means that the test ran 3 times.

Test Plan: I have adjusted the unit tests to confirm the correct behavior.

Reviewers: haoyi, cf

Reviewed By: cf

Subscribers: changesbot, mkedia

Maniphest Tasks: T57218

Differential Revision: https://tails.corp.dropbox.com/D141748