Add an api call to fetch all initial data, better login redirect
Summary: I want to add ItemOption values per user, and the best way to do that
is to return them in this api call I created. Right now I use it for auth and
the persistent message.

This also redirects the user immediately if they aren't auth'ed in.

Auditors: haoyi

Subscribers: angie