Allow tests to pull timing data from a project which may have more complete data
Summary:
Sometimes, the statistics information on test durations should not be
pulled from the timings of the current project, but some other (more
comprehensive) project.  This is particularly true for tests with
selective testing, where the parent test may not include timings for
the majority of tests that the current commit requires.

The `test_stats_from` configuration key allows the admin to provide the
name of an alternate project to pull statistics information from.

Reviewers: josiah, kylec

Reviewed By: kylec

Subscribers: changesbot, mkedia, #commit_queue

Differential Revision: https://tails.corp.dropbox.com/D141803