get_test_stats took a project object, not a project slug string
Summary:
The code added in D141803 could not work as written; the
get_test_stats function takes an object, and not a simple project slug
string.

alter the function to instead simply take the string, as only the slug
is used in the function, and a string is all that can be provided via
the configuration file.

Test Plan:
None.  Please let me know if there is a simple way to test
this.

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot, mkedia, kylec, #commit_queue

Differential Revision: https://tails.corp.dropbox.com/D145129