Apply diff builds into real local commit objects, for consistency
Summary:
By default, git apply only applies changes to the working copy. This
makes diff builds inconsistent with commit builds -- in the former,
`git diff` is sufficient to find changed files, whereas `git diff
HEAD^` is necessary in the latter. Worse, as changes are only applied
to the working directory, this makes it impossible to ascertain if new
files were added in a diff build.

Instead, make a local commit object; while `arc test`-generated builds
are in a format to simply `git am`, those generated by Phabricator and
the like are not; for simplicity and consistency, simply always use
`git apply` and an explicit `git commit` to generate the commit object.

If there exist workflows which assumed working-copy changes only (diff
builds), they will be broken by this change; however, such workflows
were already broken for all commit builds.

This is the equivalent of D145057, but for Mesos builds.

Reviewers: kylec, josiah

Reviewed By: josiah

Subscribers: changesbot, mkedia, #commit_queue

Differential Revision: https://tails.corp.dropbox.com/D145176