Added task to update the local repo.
Summary:
The task updates the repositories on locally on each machine
running a worker. Machine(s) running just Changes web should be
able to run just the worker that consumes from Celery's `repo_update`.

This removes the requirement of having a cron job that periodically
does a sync_repo to keep the repositories up-to-date on machine(s)
(the ui relies on the repositories on disk).

This is diff makes it a periodic task managed by Celery but
the task can potentially be triggered by another mechanism and
can be extended to take specific repos to refresh.

Test Plan:
Pass tests. Make sure task runs in dev environment when celery beat is on.
Test in staging environment:
There's a manually configured worker that runs on our web machine on
The reason is that it needed to run sync_repo tasks once in awhile.
Replacing that with a worker that just consumes from repo_update should work.

Reviewers: josiah, naphat, kylec

Reviewed By: kylec

Subscribers: zbarsky, changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D129625