Post coverage to Phabricator. Refactored Phabricator API.
Summary:
See https://secure.phabricator.com/T9529 for an explanation of the
Phabricator Conduit calls needed to make to post coverage to a diff.

There is one weakness in the current approach: if a developer runs
`arc diff`, starting a build, and then edits their code some more and
runs `arc diff` again before the coverage report from the first build
is posted, it's possible that the coverage from the first diff is
posted to the code of the second diff.  In order to be more precise,
we would need the phid for the exact diff (not just the revision) for
which the build was run.  I think it's possible to get this but it
would require even more Conduit calls (it already takes 4 calls to
post coverage, including the initial connect call).

I also would like to post coverage for commit builds.  This requires a
different set of Phabricator calls.  For now I dded a TODO.

Test Plan:
Wrote unit tests and looked at coverage for the new code.

I also manually ran post_diff_coverage() with dummy data and indeed it
successfully posted coverage; you can see it here:
https://tails.corp.dropbox.com/D145550?id=494164.  Also check out the
buildable from that diff: https://tails.corp.dropbox.com/B59432.

Reviewers: kylec

Reviewed By: kylec

Subscribers: kylec, changesbot, mkedia, jkao

Differential Revision: https://tails.corp.dropbox.com/D146478