Add phabricator.coverage to the list of Boolean options.
Summary:
I had missed one place where to add phabricator.coverage.  The admin
UI has a list of which options are Boolean.  Otherwise it's a string,
and the value is not coerced to either '0' or '1'.

Test Plan:
Fired up a local test instance with dummy data and observed what gets
written in the projectoption table when I flip that flag in the UI.
Without this fix, it writes 'True'.  With this fix, it writes '1'
(which is what the Python code expects).

Reviewers: kylec

Reviewed By: kylec

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D147642