Identify and handle git concurrent update errors better.
Summary:
There are cases where the failure to update a repo because of
concurrency isn't a problem or might not be a problem.
This identifies git concurrent update problems and ignores it
for jobs like sync repo and update local repo.
Failure might be problematic for things like identify_revision
in build index so we retry once. Ignoring might be reasonable too
in those cases.

We can alternatively use a file lock but this might be good enough
and that is more restrictive/less granular than the internal git lock.

Test Plan: Pass current tests.

Reviewers: kylec

Reviewed By: kylec

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D148946