post coverage to commit diffs on Phabricator revisions
Summary:
Currently we don't post coverage to Phabricator revisions after
landing when they're updated with the contents of the actual commit.  This
means that going back to look at a diff that's been landed shows all the
coverage as "Not executable", which is obviously undesirable.  This commit
causes commit builds to post coverage to their appropriate revision, if it
exists.

There doesn't appear to be a good way to get the Phabricator revision id the
commit belongs to from the Changes API, so we parse it from teh commit message
which is reasonably guaranteed to be in a regularized format.

Test Plan: ?

Reviewers: guido

Reviewed By: guido

Subscribers: changesbot, cjpark

Differential Revision: https://tails.corp.dropbox.com/D150191