Avoid notifying green builds for commit queue builds
Summary:
Commit queue builds are technically builds for a commit, but
most existing logic checks for commits (or lack of patches) to identify
builds that are for commits in the the shared global history, which commit
queue builds are not.
This restricts our green build notifications to non-commit-queue commit builds.

Test Plan: Unit

Reviewers: josiah

Reviewed By: josiah

Subscribers: nipunn, changesbot

Differential Revision: https://tails.corp.dropbox.com/D151395