Don't ask for manifest.json to be collected
Summary:
manifest.json is generated by intermediate infrastructure (namely, the generic-build script)
and exists to better associate the artifacts with the jobstep. changes-client never collects
it, and in attempting to do so, it collects unrelated files with the same name.
Additionally, when we're reporting artifacts via changes-client, the value of manifest.json
is minimal, so rather than asking for it to be collected, we'll just expect it for Jenkins
builds.

Test Plan: Existing unit

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot, anupc

Differential Revision: https://tails.corp.dropbox.com/D151955