Parse git extended diffs, detecting removal/addition of empty files
Summary:
The standard 'patch' file is not capable of expressing the addition or
removal of empty files:

    $ mkdir diff-from diff-to
    $ touch diff-from/__init__.py
    $ diff -ruN diff-from/ diff-to/ | wc -l
           0

However, these files are often significant; failing to be able to
detect a commit which simply removes or adds `__init__.py` files
causes no tests to be run, possibly at great peril.

Extend DiffParser to be able to detect git's "extended diff" style
patches; partial support for these had already been added with support
for `/dev/null` as a possible file.  This extends that support to the
file creation/deletion primitives detailed at [1].  The other directives
are not added at this time, as the information they contain is not
currently pertinent to Changes.

In so doing, the `headers` and `in_headers` variables are also removed
entirely; when the code was imported from upstream[2] in 0e5e3e9d, the
section which appended header sections was commented out.  As such,
the `headers` variable has never been used.  The `is_header`,
`new_revision`, and `old_revision` properties, which are entirely
unused, are also removed.

Finally, the diffs used in tests are updated for the current code.
This is not necessary, but does help to clarify to `git grep` that the
removal of `is_header`, for instance, is complete.

[1] https://git-scm.com/docs/git-diff#_generating_patches_with_p
[2] https://github.com/openstack-infra/lodgeit/blob/master/lodgeit/lib/diff.py

Test Plan: Added and ran tests.

Reviewers: josiah, guido

Reviewed By: guido

Subscribers: changesbot, guido, #commit_queue

Differential Revision: https://tails.corp.dropbox.com/D153599