Handle out of range test durations sanely
Summary:
We dropped some test results this weekend because the XML was reporting ~27 days of execution time.
test is our biggest table, and no tests we'll ever run will take that long, so rather than switching
to a larger type, this change truncates absurdly large and negative durations to 0 and gives us a warning.
In the future, we may want to surface a warning for the project/build, but for now, this gives us enough
info to contact the owner about it without having weird db errors or missing test failures.

Test Plan: Light unit

Reviewers: paulruan

Reviewed By: paulruan

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D153716