Expose more timestamps in build page (build start and finish, and jobstep create, start, and finish times)
Summary:
This makes it easier to see some of the overhead introduced by Changes (for example,
differences in when a jobstep is created and when it starts). They might not be
too useful for understanding where the time went without understanding exactly when
these timestamps are set but they should help narrow the search for inefficiencies
in the system, and make it a bit less mysterious for someone looking into why
a build took as long as it did.

Test Plan:
What it looks like
https://www.dropbox.com/s/4douucqf5y82fq6/Screenshot%202015-11-17%2017.03.13.png?dl=0&preview=Screenshot+2015-11-17+17.03.13.png
https://www.dropbox.com/s/lc078rw2aqsrdmn/Screenshot%202015-11-17%2016.54.25.png?dl=0

Reviewers: kylec, haoyi

Reviewed By: kylec

Subscribers: changesbot, cordelia, tabbott

Differential Revision: https://tails.corp.dropbox.com/D154319