Use the `relativePath` rather than `fileName`  as the artifact's `name`.
Summary:
This is so we can have artifacts with the same file name.
This diff adds a warning too to detect duplicate file name cases
so we can see if this is has been an actual issue.

This diff does not fix the case where there are phase results yet; in
that case, some files might still be filtered out.

Test Plan:
Fixed tests. Modified result syncing test to include
artifacts with same file names.

Reviewers: anupc, nate

Reviewed By: nate

Subscribers: changesbot, mkedia

Differential Revision: https://tails.corp.dropbox.com/D140823