When possible, extract set of changed files from VCS, not from diff
Summary:
Instead of extracting a diff from the VCS, and parsing that (which is
a computationally-intensive task for large diffs), simply directly ask
the VCS for the set of changed files.  This significantly reduces the
time to create new builds for large commit diffs.  Notably, this
allows tests for such to run via the commit queue, which is currently
the only way of testing them short of doing so on master.

Test Plan: Ran and modified tests.

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot, kylec, msolo

Differential Revision: https://tails.corp.dropbox.com/D156370