Add ability to blacklist Jenkins masters temporarily.
Summary:
- add endpoints that updates and reads a blacklist stored in redis
- add checkboxes to the jenkins by plan view for managing the blacklist

Maybe changing the checkboxes should be an admin only thing but the jenkins by plan view already contains a lot of the relevant info. Maybe we want an admin version of the same thing with the added ability of changing the blacklist?

Test Plan:
Added tests for the new endpoint.
Manually tested UI change:
https://www.dropbox.com/s/wsj9t6sejg4bbfe/Screenshot%202015-11-25%2001.05.45.png?dl=0

Reviewers: kylec, haoyi

Reviewed By: kylec

Subscribers: changesbot, kylec

Differential Revision: https://tails.corp.dropbox.com/D156213