Add UI for displaying recent infra failures
Summary:
Fixes T70914
It's useful to be able to see what has failed recently to understand current issues and to
discover new ones. This adds a view to expose (for admins) jobs that are infra failures across
all projects.
There is more that could go here (really, the longer-term goal is to have a page with a variety of
admin-targeted tools) but this is enough to be useful.

{F281523}

Test Plan: Manual

Reviewers: paulruan, haoyi

Reviewed By: paulruan

Subscribers: changesbot

Maniphest Tasks: T70914

Differential Revision: https://tails.corp.dropbox.com/D156660