Add a link to the admin page for logged-in users
Summary:
Turns out, many v2 UI users didn't think we had an admin UI because they couldn't find the link.
I played with making the link not work and show a "sorry" message for non-admins, but we should have the admin
page handle that, or in a pinch, if it is a problem, we can just only render this link for admins.

{F284784}

Test Plan: Manual

Reviewers: haoyi, paulruan

Reviewed By: paulruan

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D158204