Make jobstep_details provide the LXC config where appropriate
Summary:
Include the majority of the information required by changes-client to create and use an
LXC container in the jobstep_details JSON so we don't need to provide it on the command line
and so other tools (joyride) can reproduce the setup.

Test Plan:
Unit; will make changes-client compare to make sure we don't have
subtle differences.

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D158957