Skip snapshot builds when calculating last (passing) builds
Summary:
The "last passing build" endpoint is used to pull timing data to
allocate tests to shards.  Snapshot builds, however, are both "builds"
and "passing," but run no tests.  This causes them to be accepted as
the source of timing data, but generate sharding plans that allocate
test files randomly across all shards.

Skip snapshots when reporting last builds and last passing builds.
This causes them to never be chosen as the source of timing data.
While this does hide them from other consumers of the "last build"
endpoint, snapshot builds are a side effect of how snapshotting works,
and arguably should never be considered "builds."  As such, this
change is only likely to improve the signal that any other consumers
get from the endpoint.

Test Plan: Untested; relying on the diff build.

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot, kylec

Differential Revision: https://tails.corp.dropbox.com/D162498