Display the retry count on the single test info page
Summary:
Retries are important to know about, and we'd like to link to the test page in places where we
complain about retries, so we should show them.
However, 0 retries isn't useful, so retries are only shown if non-zero and they're
shown in red to make it clear they're bad.

Before: {F303355}
After: {F303356}

Test Plan: Manual

Reviewers: haoyi, treaster

Reviewed By: treaster

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D164247