On /commit_source/[source] page, request the maximum number of build IDs passible (100) to "ensure" that we get the build ID requested in the URL.
Summary:
On /commit_source/[source] page, request the maximum number of build IDs
passible (100) to "ensure" that we get the build ID requested in the URL.
This works around a problem where a requested build ID isn't on the first page
and results in a Javascript error, breaking the page.

Updates T76617

Test Plan: Tested manually.

Reviewers: haoyi, kylec

Reviewed By: kylec

Subscribers: changesbot, stevek

Differential Revision: https://tails.corp.dropbox.com/D166708