Drop reporting of `diff-id` for flaky tests with no first diff-id
Summary:
The `diff-id` returned from the flaky test endpoint is meant to be the
numerical diff that a test first appeared in; it does this by
examining the contents of the `target` column, and removing the first
character.

However, in the case of builds triggered by `arc test`, the `target`
column is supplied by the client, and is not always a diff-id; in the
cases where there was no existent diff yet, `target` is given as
simply the local value of `HEAD`.

Needless to say, when the first character of this (often local) commit
hash is removed, one does not find a numeric diff-id.  This causes
errors in the quarantine keeper, which attempts to update the diff
with that number.

Look up the `source` to examine if the cause of the build is actually
a Phabricator diff or not, and use the guaranteed-numeric
`revision_id` from that table as the source of truth.

This will resolve:
https://sentry.build.itc.dropbox.com/devtools/commit-queue/group/254520/

Test Plan: Fundamentally untested; I have no Changes VM to test in.

Reviewers: kylec

Reviewed By: kylec

Subscribers: josiah, changesbot, #commit_queue

Differential Revision: https://tails.corp.dropbox.com/D175672