Link to the test on the test history page instead of the build
Summary:
When this was written, we didn't have a test-specific page, but now we do, and
it seems like the appropriate thing to link to.
This also adds a helper to ChangesLinks, and uses it for the other existing creation
of this type of link.

No rendering changes.

Test Plan: Manual

Reviewers: treaster

Reviewed By: treaster

Subscribers: changesbot

Projects: #changes_ui

Differential Revision: https://tails.corp.dropbox.com/D176532