Count 'pending_allocation' as a waiting state in the UI
Summary:
We use whether a build is waiting or not to determine how jobsteps/builds
should be rendered (whether duration/results should be shown, etc).
This includes 'pending_allocation' (the state that non-Jenkins jobsteps are in
prior to having a slave assigned) as a waiting state, which should fix common
quirks like duration being rendered as "<1s" briefly while the jobstep/build is
starting out.

Test Plan: Guesswork and a heart filled with hope

Reviewers: treaster

Reviewed By: treaster

Subscribers: changesbot

Projects: #changes_ui

Differential Revision: https://tails.corp.dropbox.com/D177340