Don't offer build plan implementation options that aren't needed
Summary:
The options are hard enough to explain as is, we don't need to include unused and
potentially dangerous options in the UI.
I've verified that nobody in prod uses them. There is 1 use of DefaultBuildStep
in staging, but I'm going to kill it before this goes in.

Test Plan: Manual

Reviewers: treaster

Reviewed By: treaster

Subscribers: changesbot

Projects: #changes_ui

Differential Revision: https://tails.corp.dropbox.com/D179232