Add DummyLogFileHandler to let us fetch Jenkins logs
Summary:
I'm not sure about the use of debug_config (this is a wild-eyed backbend of a change),
but this should allow us to try out pretending that we need the contents of .log files so
they are fetched fairly easily and without making our higher-level logic more complicated.

Fixes T80928

Test Plan: Unit, and we can just try it with the xplat build.

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot

Maniphest Tasks: T80928

Differential Revision: https://tails.corp.dropbox.com/D183067