Add mypy run to the CI scripts.
Summary:
Stolen bit-for-bit from the changes-mesos-framework repo, but added `--stats` so we can see
what mypy is seeing.
`--stats` will be annoying when we get real failures because of how the runner script works,
but I think it is worth it for now.

Test Plan: Ran it, it found missing annotations. Made an incorrect annotation, and I got a useful message.

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D183093