/jobstep_summary page to display aggregate active jobstep information by cluster/project and globally.
Summary:
Fixes T87094
This is more complicated than it needs to be due to the column sorting, but I think it is worth it,
and parts of that might be breaking out for reuse.

The UI could certainly use more polish, but this is enough to be useful and pleasant.

{F368215}

Test Plan: Manual

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot

Projects: #changes_ui

Maniphest Tasks: T87094

Differential Revision: https://tails.corp.dropbox.com/D186443