Commit frequently during flake analysis to avoid holding locks too long
Summary:
Fixes T82148

I'm not confident this actually resolves the issue, and it might be safer to just do it
per project, but I think this has the potential to actually solve the problem.
Independent of this, we should switch to the batch reporting endpoint rather than
posting to dropbox.com from an open source project.

Test Plan: Unit

Reviewers: paulruan

Reviewed By: paulruan

Subscribers: changesbot

Maniphest Tasks: T82148

Differential Revision: https://tails.corp.dropbox.com/D185727