Count build results in Vortex
Summary:
This is necessarily much simpler than the data that goes to
analytics/Hive, but it gets us (a) low latency and (b) the
very nice feature of being able to set alerts.

Just don't try to read much out of stats aggregated on a
timescale less than perhaps a day.

Test Plan: unit test!

Reviewers: kylec

Reviewed By: kylec

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D187882