Version-qualify pip and easy_install invocations
Summary:
With the mypy scripts we're mixing 2.7 and 3 tools in our setup environment, which
causes confusion and is why we haven't had a successful snapshot run in a while.

Test Plan:
Hasn't been verified to fix the issue, but doesn't make it worse and is easier
to verify with a commit.

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D188673