Only email about a failed commit if it is the initial commit build
Summary:
The UI describes the field as "Addresses to notify of failures from commits (not patches)", which
suggests that addresses in that field should only be emailed in response to a failed commit build,
not mailed whenever any build based exclusively on a commit fails.
There's certainly a case for notifying in those scenarios, but moving to only mailing
for the initial commit build seems a closer match to the intended behavior.

Test Plan: Unit

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D188892