Fail cleanly when a xunit file is missing required properties
Summary:
It is dangerous for us to swallow malformed JUnit files without failing the build, so this
drops the tests and adds a FailureReason if a 'testcase' is missing 'name', which is our only
current strong requirement.
We may want to extend this to duration (we have some builds with a history of providing nonsensical
durations) or maybe use the xsd to validate, but that would certainly break some builds that currently
work and give correct pass/fail results, so limiting to only 'name' for now.

Fixes DEVTOOLS-27

Test Plan: Unit

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D189211