Remove serialize=False from API endpoints that provably need serialization
Summary:
serialize=False is an optimization for when the result is known to be JSON-able and so
running it through crumblers is unnecessary. When it turns out to not be JSON-able, we fall
back and run the crumblers anyway to avoid breaking where we can clearly respond.
These three API endpoints are on record in our logs as needing the fallback behavior, and
from reading the code it isn't hard to see where they're failing to serialize.
By removing the `serialize=False`, they should be strictly faster and a source of fewer
warnings.

Test Plan: Manual

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D189269