Speed up JobDetailsAPIView.get by making LogSource query use step_id
Summary:
The job details API can be quite slow, generally multiple seconds in prod, and almost all of that
time is spent querying for the appropriate LogSources. LogSource isn't indexed on job_id, so the
query was restricting to the current project (fairly weak) then doing a linear scan.
This adds an additional jobstep id filter subquery which allows the step_id index to be used and
in my tests reduces the time of the query from 2.6s to <20ms.

Test Plan: Manual, tests ran.

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D190609