Join Source by id in ProjectDetailsAPIView._get_avg_duration
Summary:
As the `get_any_commit_build_filters` say, the query needs to be on a join of `Build` and
`Source`. That wasn't the case for the averages query (which, by the way, we don't render the result of),
so instead of joining, it was apparently doing a linear scan on `Source` for each build.
Based on testing with manual querying on prod, this should move it from too long to reasonably run
to around 15ms (for servergit).

Test Plan: Manually ran the updated query.

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D191508