Fixing JS error when new user has no commits
Summary:
When someone tries to open the home page but does not have any commits, they currently encounter a js error because it tries to take the .length of a null type. So, first check that the list of commits actually exists.

Also, move the header up so we can show it even if they do not have commits.

Test Plan:
Return None instead of the list fetched and view in the UI.
Before:
{F384452}
After:
{F384453}

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot, kylec

Projects: #changes_ui

Differential Revision: https://tails.corp.dropbox.com/D191991