Raise exception when we fail to grab a redis lock.
Summary:
I might be missing something but it seems odd that we don't raise an error
when we fail to grab a lock and consequently not run the wrapped function.
We wrap all our tracked tasks with this lock function so failing to grab
the lock will appear as if the task finished successfully.

Reviewers: nate, josiah

Reviewed By: nate, josiah

Subscribers: changesbot, kylec

Differential Revision: https://tails.corp.dropbox.com/D192147