Filter Sources by repository in ProjectBuildIndexAPIView.get
Summary:
This makes Postgres filter Sources to those relevant to the current repository (and thus project),
which (for 'servergit') roughly reduces the set of Sources by ~2/3rds, and with that and some
other index-related behavior this triggers in the planner, the query times are observed to
go from ~4-11s to ~2.5-4s, roughly in half.
Still slow, but more reasonable.

Test Plan: Manual, both in local Changes pointed at prod and running queries against Changes prod db.

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D192304