Provide more context to Sentry when xunit parsing fails
Summary:
Primarily, this includes the content size in the warning message, which is useful in seeing whether we're
seeing truncation, and by how much.
Additionally, this'll put the actual content in the locals of the stack trace (presumably truncated in Sentry)
which will allow us to see, if truncated, whether it is the actual content.

Test Plan: None

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D193831