Silence some harmless 'unused variable' lint errors
Summary:
The 'destructing bind to grab the props we need, but also to extract
those that aren't intended for the next level so we can forward the remaining
props on' pattern is annoying for the unused var checker, but it seems to be
be perfectly idiomatic. Maybe we want to do something different, but seems fine
to just mute those lint warnings for now.
Also dropped another kinda similar bind-and-discard lint error, just because
it didn't seem worth another diff.

Test Plan: eslint

Reviewers: treaster

Reviewed By: treaster

Subscribers: changesbot

Projects: #changes_ui

Differential Revision: https://tails.corp.dropbox.com/D196513