Make coverage optional in the Source APIs.
Summary:
The source-fetching API endpoints (both of them, mostly copy-pasted) try to determine the
aggregate coverage for all builds that had run on that code. As of right now, we don't render
that anywhere, but it does have a cost linear on the number of builds that adds up to >5s in some
non-pathological cases.
It makes sense for this endpoint to provide coverage, so this makes it optional based on a paramter
to allow uses that don't care about coverage to not pay for it.

Test Plan: None

Reviewers: paulruan

Reviewed By: paulruan

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D198765