Fix sqlalchemy relations between source, repository and revision.
Summary:
Followed fix from
http://docs.sqlalchemy.org/en/latest/orm/join_conditions.html#overlapping-foreign-keys
This is currently generating sql alchemy warnings.

The way I understand the problem is that source references a repository by repository id
and a revision by sha and repository id. The existing relationships mean that both
the revision and repository can be used to change repository id.
This isn't an issue in Changes because we shouldn't ever have conflicting repository ids.
This change makes it so that changing the revision repo id won't affect source's repository id.

Test Plan: Pass tests. Not sure if it'll fix warning.

Reviewers: anupc, kylec

Reviewed By: kylec

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D202870