Redirect the LogSource to use artifactstore
Summary:
Remove saving LogChunks entirely.

Logs are fetched from Jenkins and are stored as `jenkins-console` logs in artifactstore. If changes-client is running, they will send additional logs (`console` and `infralog`) directly to the artifactstore. These logs should not conflict and should all be present on the build page. Note that logs from Jenkins are stored in a special bucket, so we need to check for it when fetching the log.

Test Plan: Run on staging with Jenkins builds. This already works locally.

Reviewers: paulruan

Reviewed By: paulruan

Subscribers: changesbot, kylec

Differential Revision: https://tails.corp.dropbox.com/D206027