Add expat parser for junit.xml files
Summary:
Switch to expat to allow ourselves to easily access Parser.CurrentByteOffset, which can be used to store messages (to be implemented in a future diff).

This diff runs both parsers together and compares the results (preferring the original one if there is a conflict). We should remove this after a short trail period.

Test Plan: Ensure behavior matches.

Reviewers: paulruan

Reviewed By: paulruan

Subscribers: changesbot, kylec

Differential Revision: https://tails.corp.dropbox.com/D206925