Add git patch id for git builds to pal_changes_build
Summary:
My end goals is to have the hash of the patch (git patch-id) available in pal_changes_build where possible. When not possible (i.e. Mercurial revision) it should be NULL.

I added a method to all Vcs classes to return this value, though the Mercurial one returns None. It is called when the revision is created.

Test Plan: Unittests

Reviewers: josiah, kylec, paulruan

Reviewed By: paulruan

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D204626