Allow a PATCH_BASE_URI to be specified for alternate patch uris.
Summary:
This is a rather coarse mechanism, but by making the patch uri global and
only available to be set from the config, there's a minimum of variation in configuration
between environments and very little potential for unapproved sneakiness.
Internally, the restricted API service for patches is available in all environments, so
this should work, and be potentially safer than what we're currently doing.

Test Plan: Unit, mypy

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D211987