Don't specify a node to avoid if the replaced jobstep was never assigned one
Summary:
We handle builds that time out while waiting for allocation poorly, because we attempt to retry
them and fail, eventually killing the `sync_job` task and leaving the build permanently hung.
This avoids one known error that can cause that.

Test Plan: The unit tests for this stuff scare me a little, but I should probably add one.

Reviewers: nate

Reviewed By: nate

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D214418