Use a fixed timeout (of 3 hours) for jobsteps in the queue/pending allocation.
Summary:
After an extended Changes outage, we tend to see a barrage of new builds on Changes, where a large
portion of jobsteps time out before getting allocated. This is especially true of jobsteps with
smaller timeouts (say, 5 mins), which are unlikely to get scheduled faster than a jobstep with
longer timeouts.

To prevent such timeouts, set a reasonably-large, fixed timeout for steps still in the queue. This
was, we allow Changes to maintain a larger queue before timing builds out.

Reviewers: josiah

Reviewed By: josiah

Subscribers: changesbot, alexmv, kylec, treaster

Differential Revision: https://tails.corp.dropbox.com/D216504