Report Status.allocated in jobstep_aggregate_by_status
Summary:
`allocated` is a state we rarely leave jobsteps in for long, but in some cases
they can be stuck there, so it'd be nice to be able to track them.
Considering how ephemeral `allocated` states typically are, I don't expect this
to have a major impact on the performance of the queries or API here.

Test Plan: Unit

Reviewers: treaster

Reviewed By: treaster

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D217879