Report time until first execution of a Task
Summary:
This Task.date_created and Task.date_modified to identify the initial run of a Task,
determine the lag from Task creation to initial execution, and report that to statsd
on a per-Task basis. Some tasks are started with a delay, but the delay (or lack of it)
is typically consistent among tasks, and we start tasks frequently enough that
this initial run delay can be a useful thing to monitor.

Test Plan: Weird unit test

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D219787