Query for the current task correctly
Summary:
We don't really make use of `Task.id`, we use `Task.task_id`, which is really `child_id`.
The original logic was broken due to this conclusion; this updates to query for the actual
primary key, and adjusts the tests accordingly.
Once this works, the additional logging will be removed.

Test Plan: Unit

Reviewers: anupc

Reviewed By: anupc

Subscribers: changesbot

Differential Revision: https://tails.corp.dropbox.com/D220573